requirements.txt
Flask
Pillow
tensorflow-cpu
virtualenv
gunicorn